+++ src/test/java/randoop/types/ClassOrInterfaceTypeTest.java
@@ -1,0 +1,103 @@
+package randoop.types;
+
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
+
+import java.util.ArrayList;
+import java.util.Collection;
+import org.junit.Test;
+import org.plumelib.util.CollectionsPlume;
+
+public class ClassOrInterfaceTypeTest {
+
+  // protected @Nullable ClassOrInterfaceType enclosingType = null;
+  // public static ClassOrInterfaceType forClass(Class<?> classType) {
+  // public static ClassOrInterfaceType forType(java.lang.reflect.Type type) {
+  // public boolean equals(@Nullable Object obj) {
+  // public int hashCode() {
+  // public abstract ClassOrInterfaceType substitute(Substitution substitution);
+  // public abstract ClassOrInterfaceType applyCaptureConversion();
+  // public String getSimpleName() {
+  // public @Nullable String getCanonicalName() {
+  // public @Nullable String getFqName() {
+  // public String getBinaryName() {
+  // public String getUnqualifiedBinaryName() {
+  // public abstract List<ClassOrInterfaceType> getInterfaces();
+  // public @Nullable Package getPackage() {
+  // public abstract NonParameterizedType getRawtype();
+  // public @Nullable InstantiatedType getMatchingSupertype(GenericClassType goalType) {
+  // public @Nullable Substitution getInstantiatingSubstitution(ReferenceType goalType) {
+  // public abstract ClassOrInterfaceType getSuperclass();
+
+  // public Collection<ClassOrInterfaceType> getSupertypes() {
+
+  @Test
+  public void getSupertypesTest() {
+    ClassOrInterfaceType cType = ClassOrInterfaceType.forClass(C.class);
+
+    Collection<ClassOrInterfaceType> cSupersStrict = cType.getSupertypesStrict();
+    if (cSupersStrict.size() != 6) {
+      throw new Error("Expected size 6: " + cSupersStrict);
+    }
+    assertTrue(CollectionsPlume.hasNoDuplicates(new ArrayList<>(cSupersStrict)));
+
+    Collection<ClassOrInterfaceType> cSupersNonstrict = cType.getSupertypesInclusive();
+    if (cSupersNonstrict.size() != 7) {
+      throw new Error("Expected size 7: " + cSupersNonstrict);
+    }
+    assertTrue(CollectionsPlume.hasNoDuplicates(new ArrayList<>(cSupersNonstrict)));
+
+    ClassOrInterfaceType dType = ClassOrInterfaceType.forClass(D.class);
+
+    Collection<ClassOrInterfaceType> dSupersStrict = dType.getSupertypesStrict();
+    if (dSupersStrict.size() != 7) {
+      throw new Error("Expected size 7: " + dSupersStrict);
+    }
+    assertTrue(CollectionsPlume.hasNoDuplicates(new ArrayList<>(dSupersStrict)));
+
+    Collection<ClassOrInterfaceType> dSupersNonstrict = dType.getSupertypesInclusive();
+    if (dSupersNonstrict.size() != 8) {
+      throw new Error("Expected size 8: " + dSupersNonstrict);
+    }
+    assertTrue(CollectionsPlume.hasNoDuplicates(new ArrayList<>(dSupersNonstrict)));
+
+    ClassOrInterfaceType objectType = ClassOrInterfaceType.forClass(Object.class);
+
+    assertTrue(objectType.getSupertypesStrict().isEmpty());
+    assertEquals(1, objectType.getSupertypesInclusive().size());
+  }
+
+  // public List<ClassOrInterfaceType> getImmediateSupertypes() {
+  // public Collection<ClassOrInterfaceType> getSupertypesInclusive() {
+  // public abstract boolean isAbstract();
+  // public boolean isGeneric(boolean ignoreWildcards) {
+  // public boolean isInstantiationOf(ReferenceType otherType) {
+  // public final boolean isNestedClass() {
+  // public final boolean isMemberClass() {
+  // public boolean isParameterized() {
+  // public abstract boolean isStatic();
+  // public boolean isSubtypeOf(Type otherType) {
+  // public boolean hasWildcard() {
+  // public boolean hasCaptureVariable() {
+  // public List<TypeArgument> getTypeArguments() {
+  // public List<TypeVariable> getTypeParameters() {
+  // public boolean isClassOrInterfaceType() {
+
+  // ///////////////////////////////////////////////////////////////////////////
+  // Classes used within tests
+  //
+
+  interface I {}
+
+  interface J extends I {}
+
+  interface K extends I {}
+
+  class A {}
+
+  class B extends A implements J {}
+
+  class C extends B implements K {}
+
+  class D extends C {}
+}