+++ src/docs/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Version ?? (?? ??, 2025)
 
+Randoop compiles and runs under Java 8-25.
+
 New command-line argument:
 
 * `--cast-to-run-time-type` enables more methods to be called on a value
@@ -649,8 +651,8 @@ and Randoop may be generating sub-optimal test suites.
 
 Fix bugs, notably:
 
-* #18 Handle non-accessible thrown exceptions
-* #51 Use better variable names in generated tests
+* Bug #18 Handle non-accessible thrown exceptions
+* Bug #51 Use better variable names in generated tests
 * Use equals, not ==, to compare NaN
 
 Many documentation improvements, including clarifications and